Rename jaxsim.utils.hashless to jaxsim.utils.wrappers
