+++ src/jaxsim/utils/__init__.py
@@ -1,5 +1,5 @@
 from jax_dataclasses._copy_and_mutate import _Mutability as Mutability
 
-from .hashless import HashlessObject
 from .jaxsim_dataclass import JaxsimDataclass
 from .tracing import not_tracing, tracing
+from .wrappers import HashlessObject
